Extract ghost cell into contrib/ghostCell module
Move all ghost cell state, config, actions, and logic out of
PositronNotebookInstance into a self-contained GhostCellController
contribution, following the same pattern as contrib/find. This keeps
the notebook instance focused on core notebook concerns and makes the
ghost cell feature independently maintainable.
