+++ src/vs/workbench/contrib/positronNotebook/browser/contrib/ghostCell/positronNotebookGhostCell.contribution.ts
@@ -1,0 +1,14 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (C) 2025 Posit Software, PBC. All rights reserved.
+ *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+// Register ghost cell actions
+import './actions.js';
+// Register ghost cell configuration
+import './config.js';
+
+import { registerPositronNotebookContribution } from '../../positronNotebookExtensions.js';
+import { GhostCellController } from './controller.js';
+
+registerPositronNotebookContribution(GhostCellController.ID, GhostCellController);
