+++ src/vs/workbench/contrib/positronNotebook/browser/positronNotebook.contribution.ts
@@ -6,6 +6,7 @@
 // Notebook editor extensions
 import './contrib/find/positronNotebookFind.contribution.js';
 import './contrib/assistant/positronNotebookAssistant.contribution.js';
+import './contrib/ghostCell/positronNotebookGhostCell.contribution.js';
 
 import { Disposable } from '../../../../base/common/lifecycle.js';
 import { Schemas } from '../../../../base/common/network.js';
@@ -42,7 +43,7 @@ import { ContextKeyExpr } from '../../../../platform/contextkey/common/contextke
 import { INotebookEditorOptions } from '../../notebook/browser/notebookBrowser.js';
 import { POSITRON_EXECUTE_CELL_COMMAND_ID, POSITRON_NOTEBOOK_EDITOR_ID, POSITRON_NOTEBOOK_EDITOR_INPUT_ID, PositronNotebookCellActionBarLeftGroup, PositronNotebookCellOutputActionGroup, usingPositronNotebooks } from '../common/positronNotebookCommon.js';
 import { getActiveCell, SelectionState } from './selectionMachine.js';
-import { POSITRON_NOTEBOOK_CELL_CONTEXT_KEYS as CELL_CONTEXT_KEYS, POSITRON_NOTEBOOK_CELL_EDITOR_FOCUSED, POSITRON_NOTEBOOK_EDITOR_FOCUSED, POSITRON_NOTEBOOK_CELL_HAS_OUTPUTS, POSITRON_NOTEBOOK_CELL_OUTPUT_COLLAPSED, POSITRON_NOTEBOOK_GHOST_CELL_AWAITING_REQUEST } from './ContextKeysManager.js';
+import { POSITRON_NOTEBOOK_CELL_CONTEXT_KEYS as CELL_CONTEXT_KEYS, POSITRON_NOTEBOOK_CELL_EDITOR_FOCUSED, POSITRON_NOTEBOOK_EDITOR_FOCUSED, POSITRON_NOTEBOOK_CELL_HAS_OUTPUTS, POSITRON_NOTEBOOK_CELL_OUTPUT_COLLAPSED } from './ContextKeysManager.js';
 import './contrib/undoRedo/positronNotebookUndoRedo.js';
 import { registerAction2, MenuId, MenuRegistry } from '../../../../platform/actions/common/actions.js';
 import { ExecuteSelectionInConsoleAction } from './ExecuteSelectionInConsoleAction.js';
@@ -1590,45 +1591,7 @@ registerNotebookWidget({
 //#endregion Notebook Header Actions
 
 //#region Ghost Cell Actions
-
-// Request ghost cell suggestion - triggered in pull mode when awaiting-request state is active
-registerAction2(class extends NotebookAction2 {
-	constructor() {
-		super({
-			id: 'positronNotebook.requestGhostCellSuggestion',
-			title: localize2('positronNotebook.requestGhostCellSuggestion', 'Request Ghost Cell Suggestion'),
-			keybinding: {
-				when: ContextKeyExpr.and(
-					POSITRON_NOTEBOOK_EDITOR_FOCUSED,
-					POSITRON_NOTEBOOK_GHOST_CELL_AWAITING_REQUEST
-				),
-				weight: KeybindingWeight.EditorContrib,
-				primary: KeyMod.CtrlCmd | KeyMod.Shift | KeyCode.KeyG
-			}
-		});
-	}
-
-	override runNotebookAction(notebook: IPositronNotebookInstance, _accessor: ServicesAccessor) {
-		notebook.requestGhostCellSuggestion();
-	}
-});
-
-// Enable ghost cell suggestions for this notebook - clears per-notebook disable setting
-registerAction2(class extends NotebookAction2 {
-	constructor() {
-		super({
-			id: 'positronNotebook.enableGhostCellSuggestionsForNotebook',
-			title: localize2('positronNotebook.enableGhostCellSuggestionsForNotebook', 'Enable Ghost Cell Suggestions for This Notebook'),
-			f1: true,
-			category: POSITRON_NOTEBOOK_CATEGORY,
-		});
-	}
-
-	override runNotebookAction(notebook: IPositronNotebookInstance, _accessor: ServicesAccessor) {
-		notebook.enableGhostCellSuggestionsForNotebook();
-	}
-});
-
+// Ghost cell actions are now registered in contrib/ghostCell/actions.ts
 //#endregion Ghost Cell Actions
 
 // Register actions
